+++ core/src/klio_core/config/core.py
@@ -25,6 +25,14 @@ from klio_core.config import _utils as utils
 
 logger = logging.getLogger("klio")
 
+
+# path used by both klio-cli and klioexec to write and read the config when
+# running a job
+RUN_EFFECTIVE_CONFIG_FILE = "klio-job-run-effective.yaml"
+RUN_EFFECTIVE_CONFIG_PATH = os.path.join(
+    "/usr/src/app", RUN_EFFECTIVE_CONFIG_FILE
+)
+
 WORKER_DISK_TYPE_URL = (
     "compute.googleapis.com/projects/{project}/regions/{region}/"
     "diskTypes/{disk_type}"
@@ -131,7 +139,7 @@ class KlioConfig(BaseKlioConfig):
             dirname = os.path.dirname(path_or_stream)
             if not os.path.exists(dirname):
                 os.makedirs(dirname)
-            with open(path_or_stream, "w+") as stream:
+            with open(path_or_stream, "w") as stream:
                 _write(stream)
         else:
             _write(path_or_stream)