+++ docs/src/reference/lib/changelog.rst
@@ -1,6 +1,16 @@
 Changelog
 =========
 
+0.2.5 (UNRELEASED)
+------------------
+
+Changed
+*******
+
+* Changed thread limiter logging to debug.
+* Workers will look for ``klio-job-run-effective.yaml`` before dropping back to ``.effective-klio-job.yaml``.
+
+
 0.2.4 (2021-01-14)
 ------------------